+++ src/Response.cpp
@@ -60,7 +60,7 @@ std::string const & etcd::Response::error_message() const
   return _error_message;
 }
 
-int etcd::Response::index() const
+int64_t etcd::Response::index() const
 {
   return _index;
 }
@@ -110,7 +110,7 @@ std::string const & etcd::Response::key(int index) const
   return _keys[index];
 }
 
-int etcd::Response::compact_revision() const
+int64_t etcd::Response::compact_revision() const
 {
   return _compact_revision;
 }
